Train continuous time system (#126)
Train a candidate lyapunov function for johansson system 2.
Store in johansson_system2_relu_10.pt. The magnitude of the Lyapunov is very small, after normalizing the Lyapunov magnitude, this is not really a Lyapunov function.

The Lyapunov positivity constraint V(x) >= rho * |x-x*|_1 is violated a lot, causing the Lyapunov function value to be very small.