Add ReLUSecondOrderSystemGivenEquilibrium class. (#187)
When we know the dynamics is second order, we don't need to learn the update on q. Because qdot = v, this part of the dynamics doesn't depend on the data. Create a class that only use the neural network to represent the update for the second order dynamics (update for v part).