In Lyapunov function, use the L1 norm of R*(x-x*). (#208)
Now we can synthesize stabilizing controller for both pendulum and the toy second order system with an non-identity R. For second order system it is important not to use R=identity. For pendulum, the data pendulum_lyapunov4.pt and pendulum_controller4.pt records a controller that stabilize the region 0 < theta < 2pi, -5 <= thetadot <=5.